LayoutTests/fast/events/pagehide-timeout-expected.txt

CONSOLE MESSAGE: line 17: Creating timeout in pagehide
CONSOLE MESSAGE: line 5: Loaded pagehide-timeout-go-back.html, going back
CONSOLE MESSAGE: line 30: Restored page from page cache.
CONSOLE MESSAGE: line 21: Timeout created in pagehide fired
Tests that JS execution is suspended after the pagehide event is fired, so that timers created within it will not fire while the document is in the page cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
